+++ source/App/source/Common.Abstractions/Actors/IActorContext.cs
@@ -12,13 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Core.App.Common.Abstractions.Actor
+namespace Energinet.DataHub.Core.App.Common.Abstractions.Actors
 {
-    public static class ActorExtensions
+    /// <summary>
+    /// Provides an actor
+    /// </summary>
+    public interface IActorContext
     {
-        public static string AsString(this Actor actor)
-        {
-            return System.Text.Json.JsonSerializer.Serialize(actor);
-        }
+        /// <summary>
+        /// Get current actor
+        /// </summary>
+        public Actor? CurrentActor { get; set; }
+
+        /// <summary>
+        /// Get DataHub as an actor
+        /// </summary>
+        public Actor DataHub { get; }
     }
 }