+++ source/App/source/FunctionApp/Middleware/ActorMiddleware.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Abstractions.Actor;
+using Energinet.DataHub.Core.App.Common.Abstractions.Actors;
 using Energinet.DataHub.Core.App.Common.Abstractions.Identity;
 using Energinet.DataHub.Core.App.FunctionApp.Extensions;
 using Energinet.DataHub.Core.App.FunctionApp.Middleware.Helpers;